a right mess
now that i can see what this app looks like, and now that the
core technologies have been validated, i am going to layer in
the default app and router. it will be a mess for a bit.

this repo started as a library, then a cli. the ui was added in
a way that let the the cli continue to function. to simplify
the ui code and align it more with a standard dioxus project,
i am going to move the ui directory up to the top level.