tear off the bandaid
at this point, everything has been broken, but the structure of the
files in the library are the objects that need to exist at the end. some
code might still be needed was deleted, but it can be retrieved from
git.

the README is right and indicates how things will work. the first
big task is to right the foundation, rebuild the cli such that it only
takes parameters, calls functions on the model, and outputs json.

after that, the UI can be tied back into the library functions.

at the end, we'll have a cli and ui built on the same code.