model: vault
a first few functions. a new semantic is being introduced here.

vault join, create, open was a little odd as it mixed vaults and
devices, and vaults needed an on disk presence before they could
join, and partially, a secondary vault was really a device.

the updated idea is that a vault is primary or secondary. secondary
vaults will have a pre-join and post-join state. pre-join, they
are only able to join, and they remain unusable until after the
first document sync. this obviates the need for device specific
home page, which was always a hack. it also allows joining to
fail and be resumed later