partial: fleshing out user concept
some pieces of data make sense most from a user perspective.
these include:
- creating and accessing keys
- minting contact records

it also wasn't making sense to have a contact record for the local
user. the local user does need a verified device list, but for some
interim states it does not make sense. it does make sense to store
a username and device name with their respective keys. this allows
the creating of the signed contract record to come later. it's maybe
a messy data model but there is precendent, ssh keys are stored with
an optional name.