local_user: clarify concepts
There are 3 types of files for a vault owner.

- .footnote/device_key

device key exists on every device. it's the iroh address and a name.

- .footnote/id_key

this is the identity signing key and username. in general, it only
exists on a single machine, the primary device. the primary device
can write new user.json records.

- .footnote/user.json

this is a signed contact record. it has the users verified devices.
it exists on all joined machines.

primary vault has:
id_key
device_key
user.json

secondary vault, unjoined:
device_key

secondary_vault, joined:
device_key
user.json

A user would be able to replace or move a primary device. A controlled
move could have a transition record that is signed with the old key and
new key. But, given known contacts, you could force import a new record
for a given petname/contact.